CromoCoin Computing
Computer programming, or in another sense computer computing, as it can be described (machine computing), is another field far from software programming, or it can be described as writing programs. Computer computing uses other languages that we cannot call programs, they are computing. Here lies the power of the machine; When it starts up and operates fully as soon as it receives a human, semi-human, or even electronic signal, then designing computer software will be like the largest missionary work in history.